refactor: validation Layer 를 RacingConfig 로 수정
